requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pyjwt==2.8.0
httpx==0.27.0
aioredis==2.0.1
pytest==8.2.1
pytest-asyncio==0.23.6
python-dotenv==1.0.1
hvac==2.2.0
bandit==1.7.8
# For type checking and linting
mypy==1.10.0
# For static analysis (optional, for CI/CD)
# sonar-scanner-cli==5.0.1.3006